Renamed repository/mock to repository/inMemory
