+++ main.go
@@ -8,14 +8,14 @@ import (
 
 	"github.com/gorilla/mux"
 	"github.com/sfqi/library/handler"
-	"github.com/sfqi/library/repository/mock"
+	"github.com/sfqi/library/repository/inMemory"
 )
 
 func main() {
 	openLibraryUrl := os.Getenv("LIBRARY")
 	olc := openlibrary.NewClient(openLibraryUrl)
 
-	db := mock.NewDB()
+	db := inMemory.NewDB()
 	bookHandler := handler.BookHandler{
 		Db:  db,
 		Olc: olc,
